Split DSODatabaseBuilder and DynamicOctree into separate files
