+++ src/celengine/CMakeLists.txt
@@ -21,6 +21,8 @@ set(CELENGINE_SOURCES
   deepskyobj.h
   dsodb.cpp
   dsodb.h
+  dsodbbuilder.cpp
+  dsodbbuilder.h
   dsooctree.cpp
   dsooctree.h
   dsorenderer.cpp
@@ -70,6 +72,7 @@ set(CELENGINE_SOURCES
   observer.cpp
   observer.h
   octree.h
+  octreebuilder.h
   opencluster.cpp
   opencluster.h
   orbitsampler.h
